Peties améliorations + création du fichier GameManager
