+++ ProjetTronClient/www/index.html
@@ -35,7 +35,7 @@
         </div>
     </div>
     <div id="waitingRoom" class="waitingRoom">
-        <h2>En attente de suffisamment de joueurs !</h2>
+        <h2>Bienvenue, vous pouvez lancer une recherche de partie !</h2>
         <p id="infoRoom"></p>
         <p id="infoJoueur"></p>
         <input type="button" value="Chercher partie" onclick="joinGame()">
